requirements.txt
# Simple requirements for beginners
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.35i==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.35
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
